resolved bugs

+++ gps_module/gps.cpp
@@ -163,8 +163,8 @@ UbxMessage Gps::readUbxMessage() {
   if (messageLength > 0 && messageLength < MAX_MESSAGE_LENGTH) {
     printf("In here MSG Length: 0x%x\n", messageLength);
       for (int i = 0; i < messageLength; i++) {
-          uint8_t byte_data = i2c_smbus_read_byte_data(i2c_fd, DATA_STREAM_REGISTER);
-          if (byte_data == static_cast<uint8_t>(-1)) {
+          int8_t byte_data = i2c_smbus_read_byte_data(i2c_fd, DATA_STREAM_REGISTER);
+          if (byte_data == -1) {
               perror("Failed to read byte from I2C device");
               return UbxMessage();  // Return an empty message on error
           }
